Refactor option parsing and default value handling for improved clarity and consistency

+++ tools/Azure.Mcp.Tools.Quota/src/Services/Util/Usage/ComputeUsageChecker.cs
@@ -23,7 +23,7 @@ public class ComputeUsageChecker(TokenCredential credential, string subscription
                 result.Add(new UsageInfo(
                     Name: item.Name?.LocalizedValue ?? item.Name?.Value ?? string.Empty,
                     Limit: (int)item.Limit,
-                    Used: (int)item.CurrentValue,
+                    Used: item.CurrentValue,
                     Unit: item.Unit.ToString()
                 ));
             }
